+++ addOns/automation/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Ensure system environment variables take precedence over configuration variables (Issue 7000).
 
 ## [0.10.0] - 2021-12-13
 ### Changed